Fix Candidate B provider CI contract
