+++ deploy/docker/mdbase-rs-revision
@@ -1,1 +1,1 @@
-a5d10abe24f15f1179424ff0e6bc7f5f6feaa834
+e2edbc8fc42f0f95c70d4dda0482b7a23a2657d7
